Add USD transaction fees chart. Clean up code. Change data type to BigDecimal.

+++ app/models/charts/transaction_fee_btc.rb
@@ -2,6 +2,6 @@ module Charts
   class TransactionFeeBtc
     include Mongoid::Document
     field :time, type: Integer
-    field :total, type: Integer
+    field :total, type: BigDecimal
   end  
 end
